Obtain a trace of inferential replicates for a sample

Simple helper function to obtain a trace (e.g. MCMC trace) of the ordered inferential replicates for one samples. Supports either multiple features, idx, or multiple samples, samp_idx (not both). Returns a tidy data.frame for easy plotting.

getTrace(y, idx, samp_idx)

Arguments

y

a SummarizedExperiment with inferential replicates as assays infRep1 etc.

idx

the names or row numbers of the gene or transcript to plot

samp_idx

the names or column numbers of the samples to plot

Value

a data.frame with the counts along the interential replicates, possible with additional columns specifying feature or sample

Examples

y <- makeSimSwishData()
getTrace(y, "gene-1", "s1")
#>    infRep count
#> 1       1    90
#> 2       2    87
#> 3       3    75
#> 4       4    92
#> 5       5    72
#> 6       6    86
#> 7       7    74
#> 8       8    93
#> 9       9   101
#> 10     10    78
#> 11     11    80
#> 12     12    62
#> 13     13    68
#> 14     14    72
#> 15     15    90
#> 16     16    91
#> 17     17    82
#> 18     18    77
#> 19     19    91
#> 20     20    71