Scale inferential replicate counts

A helper function to scale the inferential replicates to the mean sequencing depth. The scaling takes into account a robust estimator of size factor (median ratio method is used). First, counts are corrected per row using the effective lengths (for gene counts, the average transcript lengths), then scaled per column to the geometric mean sequence depth, and finally are adjusted per-column up or down by the median ratio size factor to minimize systematic differences across samples.

scaleInfReps(
  y,
  lengthCorrect = TRUE,
  meanDepth = NULL,
  sfFun = NULL,
  minCount = 10,
  minN = 3,
  saveMeanScaled = FALSE,
  quiet = FALSE
)

Arguments

y

a SummarizedExperiment with: infReps a list of inferential replicate count matrices, counts the estimated counts matrix, and length the effective lengths matrix

lengthCorrect

whether to use effective length correction (default is TRUE)

meanDepth

(optional) user can specify a different mean sequencing depth. By default the geometric mean sequencing depth is computed

sfFun

(optional) size factors function. An alternative to the median ratio can be provided here to adjust the scaledTPM so as to remove remaining library size differences. Alternatively, one can provide a numeric vector of size factors

minCount

for internal filtering, the minimum count

minN

for internal filtering, the minimum sample size at minCount

saveMeanScaled

store the mean of scaled inferential replicates as an assay 'meanScaled'

quiet

display no messages

Value

a SummarizedExperiment with the inferential replicates as scaledTPM with library size already corrected (no need for further normalization). A column log10mean is also added which is the log10 of the mean of scaled counts across all samples and all inferential replicates.

Examples

y <- makeSimSwishData()
y <- scaleInfReps(y)